- separated parallel_for_internals.hpp and parallel_for.hpp - added the possibility to control the grain size in the static scheduling   policy of the ParallelFor(Reduce) pattern(s) - ff_node::eosnotify interface changed:    virtual void eosnotify(int id=-1) --> virtual void eosnotify(ssize_t id=-1)

+++ tests/test_multi_input3.cpp
@@ -70,7 +70,7 @@ public:
         printf("got back a task from Collector\n");
         return GO_ON;
     }
-    void eosnotify(int) {
+    void eosnotify(ssize_t) {
         lb->broadcast_task((void*)EOS);
     }
 private:
